chore: create service function to epoch-scheduler
